fix beam search
* fix beam search

When you initialize beam search all beams start the same.
So the model assign the same prob to all tokens.
So topk select the same token for all beams which are already all identicals.
So all the beams produce the same translation.

To work around this, on the first call to "choose_beams" where all beams
are initialized to the same prefix, I only sample from the distribution of the first beam.

* used masked_fill instead of for loops

+++ src/fairseq2/models/nllb/tokenizer.py
@@ -102,7 +102,7 @@ class NllbTokenizer(Tokenizer):
         batch_size: Optional[int] = None,
         device: Optional[torch.device] = None,
         pin_memory: bool = False,
-        dtype: torch.dtype = torch.int32,
+        dtype: torch.dtype = torch.int64,
         disable_parallelism: bool = False,
     ) -> TokenEncoder:
         """Create a token encoder.
